Во время инициализации недоступны все кнопки. При открытии нового файла появляется окно выбора "завершить сохранение на текущем месте и открыть файл\не открывать файл"

+++ hevc_qimage_engine.h
@@ -97,7 +97,7 @@ public:
     AVFormatContext *formatContext;	   //public
 
     //Additional metadata
-    double fps_;					//may use for save?
+    int fps_;						//may use for save?
     std::string open_file_name_;	//Имя открытого HEVC файла
     int first_keyframe_ = 0;		//по умолчанию нулевой
